+++ neurocaps/_utils/_cap_internals/_cap2statmap.py
@@ -1,8 +1,10 @@
 """Internal function for turning CAPs into NifTI Statistical Maps"""
-import warnings
 import nibabel as nib, numpy as np
 from nilearn import datasets, image
 from scipy.spatial import KDTree
+from .._logger import _logger
+
+LG  = _logger(__name__)
 
 def _cap2statmap(atlas_file, cap_vector, fwhm, knn_dict):
     atlas = nib.load(atlas_file)
@@ -32,7 +34,7 @@ def _cap2statmap(atlas_file, cap_vector, fwhm, knn_dict):
         # Get non-zero indices of the stat map
         non_zero_indices = np.array(np.where(stat_map.get_fdata() != 0)).T
         if "k" not in knn_dict:
-            warnings.warn("Defaulting to k=1 since 'k' was not specified in `knn_dict`.")
+            LG.warning("Defaulting to k=1 since 'k' was not specified in `knn_dict`.")
             k = 1
         else: k = knn_dict["k"]
 
@@ -66,8 +68,8 @@ def _get_target_indices(atlas_file, knn_dict, subcortical_indices=None):
     atlas = nib.load(atlas_file)
     # Get schaefer atlas, which projects well onto cortical surface plots
     if "resolution_mm" not in knn_dict:
-        warnings.warn("Defaulting to 1mm resolution for the Schaefer atlas since 'resolution_mm' was not specified in "
-                      "`knn_dict`.")
+        LG.warning("Defaulting to 1mm resolution for the Schaefer atlas since 'resolution_mm' was not specified in "
+                   "`knn_dict`.")
         resolution_mm = 1
     else:
         resolution_mm = knn_dict["resolution_mm"]