+++ neurocaps/_utils/_timeseriesextractor_internals/_check_confound_names.py
@@ -1,5 +1,8 @@
 """Internal Function for checking confound names"""
-import warnings
+from .._logger import _logger
+
+LG = _logger(__name__)
+
 def _check_confound_names(high_pass, specified_confound_names, n_acompcor_separate):
     if specified_confound_names is None:
         if high_pass:
@@ -25,13 +28,13 @@ def _check_confound_names(high_pass, specified_confound_names, n_acompcor_separa
         if len(confound_names) > len(check_confounds):
             removed_confounds = [element for element in confound_names if element not in check_confounds]
             if specified_confound_names:
-                warnings.warn("Since `n_acompcor_separate` has been specified, specified acompcor components in "
-                              f"`confound_names` will be disregarded and replaced with the first {n_acompcor_separate} "
-                              "components of the white matter and cerebrospinal fluid masks for each participant. "
-                              f"The following components will not be used {removed_confounds}.")
+                LG.warning("Since `n_acompcor_separate` has been specified, specified acompcor components in "
+                           f"`confound_names` will be disregarded and replaced with the first {n_acompcor_separate} "
+                           "components of the white matter and cerebrospinal fluid masks for each participant. "
+                           f"The following components will not be used {removed_confounds}.")
             confound_names = check_confounds
 
-    print("List of confound regressors that will be used during timeseries extraction if available in confound "
-          f"dataframe: {', '.join(confound_names)}.", flush=True)
+    LG.info("List of confound regressors that will be used during timeseries extraction if available in confound "
+            f"dataframe: {', '.join(confound_names)}.")
 
     return confound_names